+++ packages/xchain-litecoin/src/types/bitaps-api-types.ts
@@ -1,0 +1,5 @@
+export type TxBroadcastResponse = {
+  id: string
+  result: string
+  error: string | null
+}