+++ packages/xchain-litecoin/src/types/ledger.ts
@@ -11,6 +11,6 @@ export type LedgerTxInfoParams = Pick<TxParams, 'amount' | 'recipient'> & {
   feeRate: FeeRate
   sender: Address
   network: Network
-  nodeUrl: string
+  sochainUrl: string
   nodeApiKey: string
 }