+++ packages/xchain-litecoin/src/types/sochain-api-types.ts
@@ -1,21 +1,15 @@
 import { Network, TxHash } from '@xchainjs/xchain-client'
 
 export type AddressParams = {
-  nodeUrl: string
+  sochainUrl: string
   network: Network
   address: string
 }
 
 export type TxHashParams = {
-  nodeUrl: string
+  sochainUrl: string
   network: Network
   hash: TxHash
-}
-
-export type TxBroadcastParams = {
-  nodeUrl: string
-  network: Network
-  txHex: string
 }
 
 export interface SochainResponse<T> {